Update patch set 3
Patch Set 3: Code-Review-1

Could you add description in the commit message as to why it's needed and what (or who) would use these settings?

Also, it's says 'IPv6 settings" while all the addresses (except for one) are IPv4, so what's exactly unique here for IPv6? the answer could also be added in the commit message as an explanation. 

Thanks.

Patch-set: 3
Label: Code-Review=-1